[EDIT] 한국인 학생, 유학생 계정삭제 버튼 하나로 통합
