+++ E_Global_Zone-Web_build/src/routes/Web/Manager/Students/Foreigner.js
@@ -21,7 +21,6 @@ import { useHistory, useParams } from "react-router-dom";
 import CreateSchedule from "../../../../components/common/modal/CreateSchedule";
 import Loader from "../../../../components/common/Loader";
 import { getAdminExportForeignerSect } from "../../../../api/admin/export";
-import DeleteForeignerStudent from "../../../../components/common/modal/DeleteForeignerStudent";
 
 let i = 1601214;
 let j = 0;
@@ -78,11 +77,6 @@ export default function Foreigner() {
 		isOpen: isOpenForModify,
 		handleClose: handleCloseForModify,
 		handleOpen: handleOpenForModify,
-	} = useModal();
-	const {
-		isOpen: isOpenStdDelete,
-		handleOpen: handleOpenStdDelete,
-		handleClose: handleCloseStdDelete,
 	} = useModal();
 
 	const deptList = useSelector(selectDept);
@@ -781,9 +775,6 @@ export default function Foreigner() {
 						>
 							근로 유학생 목록 저장
 						</div>
-						<div className="delete" onClick={handleOpenStdDelete}>
-							유학생 계정 삭제
-						</div>
 					</div>
 					<Modal isOpen={addIsOpen} handleClose={handleCloseForAdd}>
 						{/* <InsertForeignerStudent handleClose={handleCloseForAdd} /> */}
@@ -816,9 +807,6 @@ export default function Foreigner() {
 					</Modal>
 					<Modal isOpen={isOpenForModify} handleClose={handleCloseForModify}>
 						<ModifyForeignerStudent currentInfo={modifyInfo} reRender={reRender} />
-					</Modal>
-					<Modal isOpen={isOpenStdDelete} handleClose={handleCloseStdDelete}>
-						<DeleteForeignerStudent reRender={reRender} handleClose={handleCloseStdDelete} />
 					</Modal>
 				</div>
 			</div>
