+++ E_Global_Zone-Web_build/src/routes/Web/Manager/Students/Students.js
@@ -4,7 +4,6 @@ import ConfirmStudent from "../../../../components/common/modal/ConfirmStudent";
 import useClick from "../../../../modules/hooks/useClick";
 import ConfirmRestriction from "../../../../components/common/modal/ConfirmRestriction";
 import ConfirmUnrestriction from "../../../../components/common/modal/ConfirmUnrestriction";
-import DeleteKoreanStudent from "../../../../components/common/modal/DeleteKoreanStudent";
 import useModal from "../../../../modules/hooks/useModal";
 import {
 	getAdminKorean,
@@ -125,11 +124,6 @@ export default function Students() {
 		isOpen: isUnrestrict,
 		handleOpen: handleOpenForUnrestrict,
 		handleClose: handleCloseForUnrestrict,
-	} = useModal();
-	const {
-		isOpen: isOpenStdDelete,
-		handleOpen: handleOpenStdDelete,
-		handleClose: handleCloseStdDelete,
 	} = useModal();
 
 	/**
@@ -473,9 +467,6 @@ export default function Students() {
 					>
 						한국인 학생 목록 저장
 					</div>
-					<div className="delete" onClick={handleOpenStdDelete}>
-						한국인 학생 계정 삭제
-					</div>
 					{/*<div*/}
 					{/*	ref={useClick(function () {*/}
 					{/*		alert("엑셀 다운");*/}
@@ -512,9 +503,6 @@ export default function Students() {
 					std_kor_name={selectedKor.std_kor_name}
 					reRender={reRender}
 				/>
-			</Modal>
-			<Modal isOpen={isOpenStdDelete} handleClose={handleCloseStdDelete}>
-				<DeleteKoreanStudent reRender={reRender} handleClose={handleCloseStdDelete} />
 			</Modal>
 		</div>
 	);
